machine_learning/requirements.txt

pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.2.0
pyyaml>=6.0.0
